backend/requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-allauth==65.14.3
drf-spectacular==0.29.0
whitenoise==6.12.0
gunicorn==25.1.0
python-decouple==3.8
asgiref==3.11.1
sqlparse==0.5.5

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.11

# ─── Celery + Redis ───────────────────────────────────────────────────────────
celery==5.6.2
redis==7.2.1
django-redis==5.4.0
django-celery-beat==2.9.0
django-timezone-field==7.2.1
kombu==5.6.2
amqp==5.3.1
billiard==4.2.4
vine==5.1.0

# ─── File handling & Reports ──────────────────────────────────────────────────
Pillow==12.1.1
openpyxl==3.1.5
et_xmlfile==2.0.0

# ─── HTTP / OAuth ─────────────────────────────────────────────────────────────
requests==2.32.5
cryptography==46.0.5
PyJWT==2.11.0

# ─── API Schema / Utils ───────────────────────────────────────────────────────
uritemplate==4.2.0
inflection==0.5.1
PyYAML==6.0.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
packaging==26.0

# ─── Date / Time ──────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
python-crontab==3.3.0
cron-descriptor==1.4.5
tzdata==2025.3
tzlocal==5.3.1
six==1.17.0

# ─── CLI (Celery deps) ────────────────────────────────────────────────────────
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.6.0

# ─── Misc ─────────────────────────────────────────────────────────────────────
certifi==2026.2.25
charset-normalizer==3.4.4
cffi==2.0.0
pycparser==3.0
idna==3.11
urllib3==2.6.3
typing_extensions==4.15.0

# ─── Testing (excluded from production Docker image) ──────────────────────────
pytest==9.0.2
pytest-django==4.12.0
factory-boy==3.3.3
Faker==40.5.1
Pygments==2.19.2
iniconfig==2.3.0
pluggy==1.6.0
pdfplumber==0.11.4
